refactor(jni): moving reply logic below getViaJNI function
